Moved tests for S3BucketPermissions under tests/permissions
This is preparation work before making the actual changes to filtering
and permissions.
